fix(windows): redefine winhl on opacity update
winhl in nightly now works by linking the highlights in a window local
namespace as opposed to directly pointing to the desired group. Since
links don't follow updates, the winhl must now be reset on each opacity
update.

This is a non-optimal solution as it now requires us to make an
vimscript call on every render but it is backwards compatible. When the
next version of NeoVim is released, we can switch to using window local
namespaces entirely.

See #123